+++ R/addVarEEclosestImage.R
@@ -44,6 +44,18 @@
 addVarEEclosestImage <- function(ee_pentads, collection, reducer, maxdiff,
                                  bands = NULL, unmask = FALSE){
 
+    lifecycle::deprecate_warn(
+        when = "0.3.0",
+        what = "ABAP::addVarEEclosestImage()",
+        with = "ABDtools::addVarEEclosestImage()",
+        details = "A new version of EEcollectionToMultiband() is now on package ABDtools (https://github.com/AfricaBirdData/ABDtools).
+        The ABAP function will be discontinued",
+        id = NULL,
+        always = FALSE,
+        env = rlang::caller_env(),
+        user_env = rlang::caller_env(2)
+    )
+
   # Get image
   if(is.character(collection)){
     ee_layer <- rgee::ee$ImageCollection(collection)