+++ R/addVarEEimage.R
@@ -35,6 +35,18 @@
 addVarEEimage <- function(ee_pentads, image, reducer,
                               bands = NULL, unmask = FALSE){
 
+    lifecycle::deprecate_warn(
+        when = "0.3.0",
+        what = "ABAP::addVarEEimage()",
+        with = "ABDtools::addVarEEimage()",
+        details = "A new version of EEcollectionToMultiband() is now on package ABDtools (https://github.com/AfricaBirdData/ABDtools).
+        The ABAP function will be discontinued",
+        id = NULL,
+        always = FALSE,
+        env = rlang::caller_env(),
+        user_env = rlang::caller_env(2)
+    )
+
   # Get image
   if(is.character(image)){
     ee_layer <- rgee::ee$Image(image)